refactor(frontend): extract helpers from TwoFactorChallenge, LivePreview, CartContext (#244)
Batch 13: Extract types, constants, and utility functions into co-located helper files.
- TwoFactorChallengeStep.tsx: extract props/types, constants, base64url helpers → two-factor-challenge-helpers.ts
- LivePreview.tsx: extract ThemeOverrides, LivePreviewProps, DeviceMode, DEVICE_CONFIG → live-preview-helpers.ts
- CartContext.tsx: extract LocalCartItemDisplay, UnifiedCart, CartContextType → cart-context-types.ts
