+++ frontend/src/components/website-builder/LivePreview.tsx
@@ -12,41 +12,11 @@ import {
   loadGoogleFont,
   getFieldSelectors,
 } from './styleOverrides';
-
-interface ThemeOverrides {
-  primary_color?: string;
-  secondary_color?: string;
-  font_heading?: string;
-  font_body?: string;
-  style?: string;
-  spacing?: string;
-  button_style?: string;
-  animation?: string;
-  shadow?: string;
-  color_mode?: string;
-  bg_color?: string;
-}
-
-interface LivePreviewProps {
-  htmlContent: string | null;
-  isLoading?: boolean;
-  activeSection?: string;
-  onSectionClick?: (sectionId: string) => void;
-  onRefresh?: () => void;
-  themeOverrides?: ThemeOverrides;
-  contentOverrides?: Record<string, Record<string, unknown>>;
-  siteName?: string;
-  faviconUrl?: string;
-  siteUrl?: string;
-}
-
-type DeviceMode = 'desktop' | 'tablet' | 'mobile';
-
-const DEVICE_CONFIG: Record<DeviceMode, { width: string; label: string; scale?: boolean }> = {
-  desktop: { width: '100%', label: 'Escritorio' },
-  tablet: { width: '768px', label: 'Tablet', scale: true },
-  mobile: { width: '375px', label: 'Móvil', scale: true },
-};
+import {
+  type LivePreviewProps,
+  type DeviceMode,
+  DEVICE_CONFIG,
+} from './live-preview-helpers';
 
 export default function LivePreview({
   htmlContent,
